refactor: rename redux-states/task/types.ts to model.ts
