+++ src/redux-states/task/index.ts
@@ -1,5 +1,5 @@
 import * as taskActions from './actions';
-import * as taskTypes from './types';
+import * as taskTypes from './model';
 import taskReducer from './reducer';
 
 export { taskActions, taskReducer, taskTypes };
