+++ src/redux-states/task/model.ts
